Placed kernel in higher memory at location 0xFFFFFFFF80100000.
Separated boot.S into two files (boot.S and load.S).  boot.S is the 32 
bit code that is linked to a physical address.  load.S is the 64 bit 
code that is the first code linked to a virtual address.

boot.h contains the constants the give where the kernel is located in 
both virtual and physical memory.

linker.ld is the linker script that the kernel is linked with which will 
offset code to the virtual address.  It is spiffy keen.

I want to call the kernel "kmain" because it sounds cooler, and I don't 
like the connotation that it is a purely C function and would rather it 
be apparent this is the main function of the kernel.  Anyone who argues, 
I will swordfight your mother...that's right, I said it, I will fence 
the woman who gave birth to you!

Some hacks were used to get the instruction pointer to the virtual 
address.  (load.S on the lretq) Just word to the wise.  This is so we 
can quickly tear apart that linear addressing and 1-1 mapping as 
quickly as possible and get virtual memory implemented.

I added comments to the files I worked on, which we should always do, 
which give a synopsis on what is going on and the responsibilites of 
each file.  Very important with the technical BS in the assembly 
portions.

The makefile is updated with comments and to add load.S to the assembly 
compile list.  Also, the linker flags are separated for kernel and 
application so that I may use the kernel linker script (linker.ld) which 
is now a part of LDFLAGS.  GDC and GCC both have the -mcmodel=kernel 
flag.  This mysterious flag solved my relocation problems (mostly) and I 
had an awful time before figuring out that one.  Apparently it does 
stuff differently when in the "kernel" memory model that it doesn't do 
any other time.

Yeah... I was bored, I suppose.  Oh well.  Got stuff done.  Over and 
out!  ~ Wilks! 

PS: Now, would everyone be this descriptive in your commits?