Well, now I'm told that config.d is not to be put in the svn.  Note: if you do a clean checkout and find yourself sans a config.d, copy config.d.sample to config.d to compile.
The point of not having config.d is to ensure everyone can have their 
own different copy.  I see this as an issue, different config.d could 
mean different environments and testing can be a problem.  It seems we 
may need to synch it every once and a while.  This can be debated later.

~ Wilks