Added capabilities to the MP code.  It will now search the correct region, and will search the BIOS ROM region first.  It will only return success when the table is verified.  The configuration table will also be taken from the MP table and used to gather the different Processor, Bus, IO Apic, IO Interrupt, and Local Interrupt entries.
The APIC register region is also exposed, but the kernel needs to have
mapped in the pages into virtual memory before they can be used.  I will
code this functionality later using the functions in vmem and use the
registers to enabled the boot strap processor's local APIC.

The MP table is only generated by bochs' bios when PCI is enabled, word
to the wise.  Now we can get to the real stuff.  Woo.

Wilks!