Got that second CPU into long mode and into the kernel.  It shares page tables with CPU 0 (the boot CPU).  It uses locks to not screw stuff up, but nothing is truly MP-safe.  I might change the naming of the locks to reflect conventions in the code and clean it up with comments, but the code itself looks pretty spiffy for it.  Nice work, Eric.  Come back for the scheduler!!!
I might have to go and reconsider how the trampoline works so that it 
can share as much code as in the boot.S and load.S.  Take care of the 
redundancy.  Also, I'm not sure about how to allocate a stack for every 
cpu I find yet.  I just made another stack space, and changed how this 
stack space is linked (fyi, not within the common, as it doesn't like 
there being more than one?).  Either way, it works like a charm.